+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add Tree View groups (#20)
 - Add an option to display the variable name in front of the value (#30)
 - Increase tests coverage (#31)
+- Add auto scroll to the selected value (#32)
 
 ## 1.3.0 (2023-06-30)